add dynamic show image

+++ app/src/main/java/com/sori/zenvo/util/DensityUtil.java
@@ -6,13 +6,13 @@ import android.content.Context;
  * Created by zhouyi on 2015/12/30.
  */
 public class DensityUtil {
-    public static int dip2px(Context context, float dpValue) {
+    public static float dip2px(Context context, float dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (dpValue * scale + 0.5f);
+        return dpValue * scale + 0.5f;
     }
 
-    public static int px2dip(Context context, float pxValue) {
+    public static float px2dip(Context context, float pxValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue / scale + 0.5f);
+        return pxValue / scale + 0.5f;
     }
 }
